NK_HW_21
removed all the Thread.sleep
