+++ src/test/java/Homework17.java
@@ -3,7 +3,7 @@ import org.testng.annotations.Test;
 public class Homework17 extends BaseTest{
 
     @Test
-    public void searchAndAddSongToPlaylist() throws InterruptedException {
+    public void searchAndAddSongToPlaylist() {
 
         provideEmail("natalia.kalinina@testpro.io");
         providePassword("nkKoel24$");
