+++ src/test/java/Homework18.java
@@ -3,7 +3,7 @@ import org.testng.annotations.Test;
 public class Homework18 extends BaseTest{
 
     @Test
-    public void playSong() throws InterruptedException {
+    public void playSong() {
 
         navigateToPage();
         provideEmail("natalia.kalinina@testpro.io");
