+++ src/test/java/Homework21.java
@@ -1,0 +1,2 @@
+public class Homework21 extends BaseTest{
+}
